+++ src/rayoptics/mpl/analysisfigure.py
@@ -9,6 +9,7 @@
 """
 
 import numpy as np
+from matplotlib.colors import PowerNorm
 
 from rayoptics.mpl.styledfigure import StyledFigure
 
@@ -16,6 +17,7 @@ from rayoptics.optical import analyses
 
 
 class AnalysisFigure(StyledFigure):
+    """Containing Figure for single panel plots, supports update_data."""
 
     def __init__(self, data_objs, **kwargs):
 
@@ -42,6 +44,24 @@ class AnalysisFigure(StyledFigure):
 
 
 class RayFanPlot():
+    """Single axis line plot, supporting data display from multiple RayFans.
+
+    Attributes:
+        fig: the parent figure for this panel
+        gs: gridspec for this panel
+        fan_list: list of (fan, data_type, kwargs)
+
+            - fan: a RayFan instance
+            - data_type: 'x', 'y', 'opd' to be extracted from fan
+            - kwargs: passed to axis.plot() call
+
+        scale_type: if 'fit', set scale to encompass largest data value
+        user_scale_value: max scale to apply if scale_type is 'user'
+        title: title, if desired, of this plot panel
+        yaxis_ticks_position: 'left' or 'right', default is 'left'
+        kwargs: passed to plot call
+
+    """
 
     def __init__(self, fig, gs, fan_list,
                  user_scale_value=0.1, scale_type='fit',
@@ -111,19 +131,41 @@ class RayFanPlot():
 
 
 class RayGeoPSF():
+    """Single axis spot diagram or 2d histogram.
+
+    Attributes:
+        fig: the parent figure for this panel
+        gs: gridspec for this panel
+        ray_list: a RayList instance
+        dsp_type: display type, either 'spot' or 'hist2d'
+        scale_type: if 'fit', set scale to encompass largest data value
+        user_scale_value: max scale to apply if scale_type is 'user'
+        title: title, if desired, of this plot panel
+        yaxis_ticks_position: 'left' or 'right', default is 'left'
+        kwargs: passed to plot call
+    """
 
     def __init__(self, fig, gs, ray_list,
                  user_scale_value=0.1, scale_type='fit',
-                 yaxis_ticks_position='left',
+                 yaxis_ticks_position='left', dsp_typ='hist2d',
                  **kwargs):
         self.fig = fig
         self.fig.subplots.append(self)
 
         self.gs = gs
         self.ray_list = ray_list
+        self.dsp_typ = dsp_typ
 
         if 'title' in kwargs:
             self.title = kwargs.pop('title', None)
+
+        if 'norm' in kwargs:
+            self.norm = kwargs.pop('norm', None)
+        else:
+            gamma = 0.3
+            vmax = kwargs.get('vmax') if 'vmax' in kwargs else None
+            self.norm = PowerNorm(gamma, vmin=0., vmax=vmax)
+
         self.plot_kwargs = kwargs
 
         self.user_scale_value = user_scale_value
@@ -133,10 +175,12 @@ class RayGeoPSF():
         self.update_data()
 
     def init_axis(self, ax):
-        ax.grid(True)
-        ax.axvline(0, c=self.fig._rgb['foreground'], lw=1)
-        ax.axhline(0, c=self.fig._rgb['foreground'], lw=1)
-        # ax.tick_params(labelbottom=False)
+        if self.dsp_typ == 'spot':
+            ax.grid(True)
+            linewidth = 0.5
+            ax.axvline(0, c=self.fig._rgb['foreground'], lw=linewidth)
+            ax.axhline(0, c=self.fig._rgb['foreground'], lw=linewidth)
+
         ax.yaxis.set_ticks_position(self.yaxis_ticks_position)
         if self.title is not None:
             ax.set_title(self.title, fontsize='small')
@@ -154,26 +198,48 @@ class RayGeoPSF():
         ax = self.fig.add_subplot(self.gs)
         self.init_axis(ax)
 
-        ax.scatter(*self.ray_list.ray_abr, **self.plot_kwargs)
-
-        ax.set_aspect('equal')
-
         if self.scale_type == 'fit':
             x_data = self.ray_list.ray_abr[0]
             y_data = self.ray_list.ray_abr[1]
             max_value = max(max(np.nanmax(x_data), -np.nanmin(x_data)),
                             max(np.nanmax(y_data), -np.nanmin(y_data)))
-            ax.set_xlim(-max_value, max_value)
-            ax.set_ylim(-max_value, max_value)
         elif self.user_scale_value is not None:
-            us = self.user_scale_value
-            ax.set_xlim(-us, us)
-            ax.set_ylim(-us, us)
+            max_value = self.user_scale_value
+
+        ax.set_xlim(-max_value, max_value)
+        ax.set_ylim(-max_value, max_value)
+
+        if self.dsp_typ == 'spot':
+            ax.scatter(*self.ray_list.ray_abr, **self.plot_kwargs)
+        elif self.dsp_typ == 'hist2d':
+            edges = np.linspace(-max_value, max_value, num=100)
+            h, xedges, yedges, qmesh = ax.hist2d(*self.ray_list.ray_abr,
+                                                 bins=edges,
+                                                 norm=self.norm,
+                                                 **self.plot_kwargs)
+            ax.set_facecolor(qmesh.cmap(0))
+
+        ax.set_aspect('equal')
 
         return self
 
 
 class Wavefront():
+    """Single axis wavefront map.
+
+    Attributes:
+        fig: the parent figure for this panel
+        gs: gridspec for this panel
+        ray_grid: a RayGrid instance
+        do_contours: if True, display contour plot, else plot data grid as an
+        image
+        scale_type: if 'fit', set scale to encompass largest data value
+        user_scale_value: max scale to apply if scale_type is 'user'
+        title: title, if desired, of this plot panel
+        yaxis_ticks_position: 'left' or 'right', default is 'left'
+        cmap: color map for plot, defaults to 'RdBu_r'
+        kwargs: passed to plot call
+    """
 
     def __init__(self, fig, gs, ray_grid,
                  do_contours=False, user_scale_value=None,
@@ -186,6 +252,9 @@ class Wavefront():
 
         if 'title' in kwargs:
             self.title = kwargs.pop('title', None)
+        kwargs['cmap'] = kwargs.get('cmap', "RdBu_r")
+        self.plot_kwargs = kwargs
+
         self.do_contours = do_contours
         self.user_scale = user_scale_value
 
@@ -219,18 +288,18 @@ class Wavefront():
             hmap = ax.contourf(grid[0],
                                grid[1],
                                grid[2],
-                               cmap="RdBu_r",
                                vmin=-max_value,
-                               vmax=max_value)
+                               vmax=max_value,
+                               **self.plot_kwargs)
         else:
             # transpose and origin=lower needed to match display orientation
             # of contours
             hmap = ax.imshow(np.transpose(grid[2]),
                              origin='lower',
-                             cmap="RdBu_r",
                              vmin=-max_value,
                              vmax=max_value,
                              extent=[-1., 1., -1., 1.],
+                             **self.plot_kwargs
                              )
         self.fig.colorbar(hmap, ax=ax, use_gridspec=True)
         ax.set_aspect('equal')